🚨 Rename routes/import/lib.js to lib.svelte.js
